fix memory leak + sc naming (#47)
